refactor: move posts pages/components into features with stubs
